+++ database/migrations/2019_09_01_214911_create_proformas_table.php
@@ -26,7 +26,7 @@ class CreateProformasTable extends Migration
             $table->string('brand')->nullable();
             $table->string('origin')->nullable();
             $table->integer('vat_amount')->nullable();
-            $table->integer('total_to_pay')->nullable();
+            $table->decimal('total_to_pay', 14, 2)->nullable();
             $table->integer('is_received')->default(0);
             $table->text('note')->nullable();
             $table->timestamps();
